+++ apps/web/client/src/app/project/[id]/_components/top-bar/publish/dropdown/custom-domain-section.tsx
@@ -29,7 +29,7 @@ export const CustomDomainSection = observer(() => {
 
     const openCustomDomain = (): void => {
         editorEngine.state.publishOpen = false;
-        editorEngine.state.settingsTab = SettingsTabValue.DOMAIN;
+        stateManager.settingsTab = SettingsTabValue.DOMAIN;
         stateManager.isSettingsModalOpen = true;
     };
